Add Downsampling and Radon to functional
