+++ deepinv/physics/functional/downsampling.py
@@ -1,0 +1,4 @@
+from torch import Tensor
+
+def downsample(x: Tensor, factor: int):
+    return x[:, :, :: factor, :: factor]
